Build abstraction layer between poeditor and translations
