+++ src/data/entities/Translation.ts
@@ -5,3 +5,14 @@ export interface Translation {
     provider: TranslationProvider;
     projectId: string;
 }
+
+export interface TranslationProject {
+    id: string;
+    name: string;
+    languages: TranslationLanguage[];
+}
+
+export interface TranslationLanguage {
+    id: string;
+    name: string;
+}
